+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2026-07-05
+
+### Changed
+- Pandas Join no accepts multiple dataframes
+- Funcflow concurrent fork no defaults to the number of workers equal to the number of callables
+
+### Added
+- Pandas and Polars interpolate
+
+
 ## [1.1.0] - 2026-06-28
 
 ### Fixed